Use Dagger to init the bot
